+++ src/mol-plugin/state.ts
@@ -89,7 +89,13 @@ class PluginState extends PluginComponent {
         if (snapshot.behaviour) await this.plugin.runTask(this.behaviors.setSnapshot(snapshot.behaviour));
         if (snapshot.data) await this.plugin.runTask(this.data.setSnapshot(snapshot.data));
         if (snapshot.canvas3d?.props) {
-            const settings = PD.normalizeParams(Canvas3DParams, snapshot.canvas3d.props, 'children');
+            const settings: Partial<Canvas3DProps> = PD.normalizeParams(Canvas3DParams, snapshot.canvas3d.props, 'children');
+            if (snapshot.camera?.current || snapshot.camera?.focus) {
+                // Avoid multiple camera transitions (creates ugly cases when camera in old and new snapshot is the same)
+                settings.camera = undefined;
+                settings.cameraClipping = undefined;
+                settings.cameraFog = undefined;
+            }
             await PluginCommands.Canvas3D.SetSettings(this.plugin, { settings });
         }
         if (snapshot.canvas3dContext?.props) {